+++ latexonhttp/utils/misc.py
@@ -7,7 +7,7 @@ LaTeX-On-HTTP miscellaneous utils.
 :copyright: (c) 2019-2022 Yoan Tournade.
 :license: AGPL, see LICENSE for more details.
 """
-CURRENT_API_VERSION = "2025-02-13-1"
+CURRENT_API_VERSION = "2025-09-23-1"
 
 
 def get_api_version():